% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_my_env.R
\name{build_my_env}
\alias{build_my_env}
\title{Create two directories: `./data` and `./output`}
\usage{
build_my_env()
}
\value{
not return
}
\description{
Create two directories: `./data` and `./output`
}
