% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{contain_all_keywords}
\alias{contain_all_keywords}
\title{contain_all_keywords}
\usage{
contain_all_keywords(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{keywords}
}
\value{
logical
}
\description{
contain_all_keywords
}
