% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{get_past_year}
\alias{get_past_year}
\title{get_past_year}
\usage{
get_past_year(.date_var, period = 1, tf = "\%Y-\%m")
}
\arguments{
\item{.date_var}{a}

\item{period}{a}

\item{tf}{a}
}
\description{
get_past_year
}
