% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinese_number.R
\name{numeric2Chinese}
\alias{numeric2Chinese}
\title{numeric to Chinese}
\usage{
numeric2Chinese(number, type = "upper")
}
\arguments{
\item{number}{integer}

\item{type}{upper or lower chinese number}
}
\value{
character
}
\description{
numeric to Chinese
}
\examples{
numeric2Chinese(123)
}
