% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt_mof_export_summary.R
\name{rpt_mof_export_summary}
\alias{rpt_mof_export_summary}
\title{mof export summary by each country and each industry}
\usage{
rpt_mof_export_summary(
  start_date,
  end_date,
  period = 3,
  sub_hs_digits = 11,
  country_list = NULL,
  money = "usd",
  industry_type = "all_industry",
  destdir = ".",
  verbose = TRUE
)
}
\arguments{
\item{start_date}{yyyy-mm}

\item{end_date}{yyyy-mm}

\item{period}{the period of cagr year}

\item{sub_hs_digits}{hscode subset}

\item{country_list}{filter specified countries}

\item{money}{usd or twd}

\item{industry_type}{"all_industry", "industry21", "version1", "version2"}

\item{destdir}{output directory}

\item{verbose}{TRUE or FALSE}
}
\value{
a list contains output file name and data
}
\description{
calculate mof export summary by each country and each industry
}
