% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{rpt_mof_industry_region}
\alias{rpt_mof_industry_region}
\title{Output report 1}
\usage{
rpt_mof_industry_region(.start_date, .end_date, .industry_keyword,
  .region = "country", .direct = "export", .money = "usd",
  .fixed_cny_nm = TRUE)
}
\arguments{
\item{.start_date}{start date}

\item{.end_date}{end date}

\item{.industry_keyword}{industry}

\item{.region}{country or area}

\item{.direct}{import or export}

\item{.money}{usd or twd}

\item{.fixed_cny_nm}{fixed country names}
}
\value{
list
}
\description{
Output report 1
}
