% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt_mof_summary_dev.R
\name{rpt_mof_summary_dev}
\alias{rpt_mof_summary_dev}
\title{rpt_mof_summary_dev}
\usage{
rpt_mof_summary_dev(start_date, end_date, period = 3,
  sub_hs_digits = 11, n = 3, country_list = NULL,
  direct = "export", money = "usd", industry_type = "all_industry",
  destdir = ".", verbose = TRUE)
}
\arguments{
\item{start_date}{.}

\item{end_date}{.}

\item{period}{.}

\item{sub_hs_digits}{.}

\item{n}{.}

\item{country_list}{.}

\item{direct}{.}

\item{money}{.}

\item{industry_type}{.}

\item{destdir}{.}

\item{verbose}{.}
}
\value{
list
}
\description{
rpt_mof_summary_dev
}
