% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{str2regex}
\alias{str2regex}
\title{str2regex}
\usage{
str2regex(x, sep = ",", sub = 100, start = "^", end = "")
}
\arguments{
\item{x}{hscode string}

\item{sep}{a}

\item{sub}{a}

\item{start}{a}

\item{end}{a}
}
\value{
string
}
\description{
str2regex
}
