% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taitratools.R
\docType{package}
\name{taitratools}
\alias{taitratools}
\alias{taitratools-package}
\title{taitratools: A package for work.}
\description{
The taitratools package provides two categories of important functions:
get source path, and read file.
}
\section{tt_read_table functions}{

The tt_read_table functions ...
}

