% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_area.R
\name{tt_bind_area}
\alias{tt_bind_area}
\title{Adding an area column to input data.frame}
\usage{
tt_bind_area(.df)
}
\arguments{
\item{.df}{A data.frame}
}
\value{
A data.frame
}
\description{
The data.frame must contain a column named `country`.
}
