% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_industry.R
\name{tt_bind_industry}
\alias{tt_bind_industry}
\title{Adding industry label to input data.frame}
\usage{
tt_bind_industry(.df, sub = 11, ind21 = FALSE, index = NULL,
  type = NULL, major = NULL, minor = NULL)
}
\arguments{
\item{.df}{A data frame. Input data.frame MUST contain `hscode` column!}

\item{sub}{The digits of `hscode`. If you set the digit less than 11, alway remember summarise the value of Input data.}

\item{ind21}{Boolean, If TRUE, output industry 21 data.}

\item{index}{Filter specified index.}

\item{type}{Filter specified type.}

\item{major}{Filter specified major.}

\item{minor}{Filter specified minor.}
}
\value{
A data frame
}
\description{
Adding industry label to input data.frame
}
