% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{tt_compare_hscode}
\alias{tt_compare_hscode}
\title{Compare two hscode is the same}
\usage{
tt_compare_hscode(x, y, sub = NULL)
}
\arguments{
\item{x}{hscode vector 1}

\item{y}{hscode vector 2}

\item{sub}{integer. The last element to be replaced.}
}
\description{
Compare two hscode is the same
}
