% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_industry.R
\name{tt_convert_industry_name}
\alias{tt_convert_industry_name}
\title{convert report_industry_name}
\usage{
tt_convert_industry_name(major, minor)
}
\arguments{
\item{major}{string}

\item{minor}{string}
}
\value{
string
}
\description{
convert report_industry_name
}
