% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_df_filter_top_and_bottom_n.R
\name{tt_df_filter_top_and_bottom_n}
\alias{tt_df_filter_top_and_bottom_n}
\title{tt_df_filter_top_and_bottom_n}
\usage{
tt_df_filter_top_and_bottom_n(x, n, wt)
}
\arguments{
\item{x}{a data.frame to filter}

\item{n}{number of row to return}

\item{wt}{variable}
}
\value{
data.frame
}
\description{
tt_df_filter_top_and_bottom_n
}
