% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_df_mutate_chinese_hscode.R
\name{tt_df_mutate_chinese_hscode}
\alias{tt_df_mutate_chinese_hscode}
\title{tt_df_mutate_chinese_hscode}
\usage{
tt_df_mutate_chinese_hscode(x, nm = "hscode_ch")
}
\arguments{
\item{x}{data.frame}

\item{nm}{chinese hscode name}
}
\value{
data.frame
}
\description{
tt_df_mutate_chinese_hscode
}
