% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_df_sub_hscode.R
\name{tt_df_sub_hscode}
\alias{tt_df_sub_hscode}
\title{tt_df_sub_hscode}
\usage{
tt_df_sub_hscode(x, start = 1L, end = -1L, by = "value",
  forct_year = TRUE)
}
\arguments{
\item{x}{data}

\item{start}{integer}

\item{end}{integer}

\item{by}{"value" or "count" or "weight"}

\item{forct_year}{drop month}
}
\value{
data.frame
}
\description{
tt_df_sub_hscode
}
