% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_columns_name.R
\name{tt_df_upper_name}
\alias{tt_df_upper_name}
\title{convert column name upper case}
\usage{
tt_df_upper_name(x)
}
\arguments{
\item{x}{data frame}
}
\value{
data frame
}
\description{
convert column name upper case
}
