% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/English2Chinese.R
\name{tt_English2Chinese}
\alias{tt_English2Chinese}
\title{Convert English name to Chinese name}
\usage{
tt_English2Chinese(x, types)
}
\arguments{
\item{x}{data.frame}

\item{types}{only support itc, un_comtrade and asean now.}
}
\value{
data.frame
}
\description{
Convert English name to Chinese name
}
