% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_world_value.R
\name{tt_ext_world_value}
\alias{tt_ext_world_value}
\title{tt_ext_world_value}
\usage{
tt_ext_world_value(x, column_nm = NULL)
}
\arguments{
\item{x}{data.frame}

\item{column_nm}{column name}
}
\value{
numeric value
}
\description{
input data, output data only world value
}
