% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_area.R
\name{tt_filter_by_area}
\alias{tt_filter_by_area}
\title{Filter data by area}
\usage{
tt_filter_by_area(.df, area)
}
\arguments{
\item{.df}{A data frame}

\item{area}{A character vector}
}
\value{
data frame
}
\description{
Filter data by area
}
