% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_hscode.R
\name{tt_filter_by_hscode}
\alias{tt_filter_by_hscode}
\title{Filter data by specific hscode}
\usage{
tt_filter_by_hscode(.df, hscode, new_col = NULL, sub = 10)
}
\arguments{
\item{.df}{A data frame}

\item{hscode}{A string or character vector}

\item{new_col}{Create new column name}

\item{sub}{hscode digits}
}
\value{
data frame
}
\description{
Filter data by specific hscode
}
