% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{tt_format_hscode}
\alias{tt_format_hscode}
\title{format hscode}
\usage{
tt_format_hscode(x, sub = NULL, start = "", end = "",
  collapse = ",", .order = TRUE)
}
\arguments{
\item{x}{hscode charactor}

\item{sub}{integer. The last element to be replaced.}

\item{start}{a character}

\item{end}{a character}

\item{collapse}{an optional character string to separate the results.}

\item{.order}{order}
}
\description{
format hscode
}
