% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{tt_genReport_1}
\alias{tt_genReport_1}
\title{Output report 1}
\usage{
tt_genReport_1(.start_date, .end_date, .direct = "export",
  .money = "usd", .industry_keyword)
}
\arguments{
\item{.start_date}{start date}

\item{.end_date}{end date}

\item{.direct}{import or export}

\item{.money}{usd or twd}

\item{.industry_keyword}{industry}
}
\value{
list
}
\description{
Output report 1
}
