% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{tt_get_path}
\alias{tt_get_path}
\title{Get all source path}
\usage{
tt_get_path(path_name)
}
\arguments{
\item{path_name}{the name of source}
}
\value{
the source's path
}
\description{
You can use `tt_ls()` to see all source and name.
}
\examples{

tt_get_path("PATH_AREA")

tt_get_path("PATH_FULL_HSCODE")

}
