% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{tt_grouped_sum}
\alias{tt_grouped_sum}
\title{grouped sum}
\usage{
tt_grouped_sum(data, ..., by = "value")
}
\arguments{
\item{data}{data frame}

\item{...}{grouped columns}

\item{by}{summarise by value or count or weight}
}
\value{
data frame
}
\description{
grouped sum
}
