% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_show.R
\name{tt_hscode_show}
\alias{tt_hscode_show}
\title{Show hscode chinese name}
\usage{
tt_hscode_show(hscode = NULL, chinese = NULL)
}
\arguments{
\item{hscode}{hscode}

\item{chinese}{chinese}
}
\value{
data frame
}
\description{
Show hscode chinese name
}
