% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industry_grouped_sum.R
\name{tt_industry_grouped_sum}
\alias{tt_industry_grouped_sum}
\title{tt_industry_grouped_sum}
\usage{
tt_industry_grouped_sum(x, industryType = "all_industry")
}
\arguments{
\item{x}{data.frame}

\item{industryType}{"all_industry", "industry21", "version1", "version2"}
}
\value{
a list
}
\description{
tt_industry_grouped_sum
}
