% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.R
\name{tt_ls}
\alias{tt_ls}
\title{List all source path}
\usage{
tt_ls()
}
\value{
Source imformation.
}
\description{
This is a helper funciton that can list all source path on screen.
The output has three part, first part is status, if files or directories are not exist,
it would show a smile or bad face. The second part is source name, and last part is the source path.
}
\examples{

tt_ls()

}
