% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_by_industry_for_reports.R
\name{tt_order_by_industry_for_reports}
\alias{tt_order_by_industry_for_reports}
\title{order by industry for reports}
\usage{
tt_order_by_industry_for_reports(df, version = NULL)
}
\arguments{
\item{df}{data.frame}

\item{version}{version 1 or version 2}
}
\value{
data.frame
}
\description{
order by industry for reports
}
