% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{tt_parse_date}
\alias{tt_parse_date}
\title{Parse_date
Parse start date and end date}
\usage{
tt_parse_date(start_date, end_date, period = NA)
}
\arguments{
\item{start_date}{yyyy-mm}

\item{end_date}{yyyy-mm}

\item{period}{integer}
}
\value{
a list
}
\description{
Parse_date
Parse start date and end date
}
