% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_parse_hscode.R
\name{tt_parse_hscode}
\alias{tt_parse_hscode}
\title{tt_parse_hscode}
\usage{
tt_parse_hscode(x)
}
\arguments{
\item{x}{a character of hscode}
}
\value{
a character of hscode chinese name
}
\description{
tt_parse_hscode
}
