% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{tt_read_table}
\alias{tt_read_table}
\title{Easily read files.}
\usage{
tt_read_table(path, col_types = NULL, sheet = NULL)
}
\arguments{
\item{path}{string}

\item{col_types}{column type}

\item{sheet}{integer}
}
\value{
data.frame
}
\description{
This function is based on readr, readxls, and feather packages.
You can easily read a file just with one function.
Support file format: `.csv`, `.tsv`, `.xlsx`, `.xls`, `.feather`, `.rds`.
}
