% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_value.R
\name{tt_spread_value}
\alias{tt_spread_value}
\title{spread data by specified column}
\usage{
tt_spread_value(x, by, na2zero = TRUE)
}
\arguments{
\item{x}{a data.frame}

\item{by}{a character}

\item{na2zero}{replace value with zero or not.}
}
\value{
data.frame
}
\description{
spread data by specified column
}
