% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_world_value.R
\name{tt_ext_world_value}
\alias{tt_ext_world_value}
\title{tt_ext_world_value}
\usage{
tt_ext_world_value(x, column_nm = NULL, by_country = FALSE)
}
\arguments{
\item{x}{data.frame}

\item{column_nm}{column name}

\item{by_country}{by country or not}
}
\value{
numeric value
}
\description{
input data, output data only world value
}
