% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{tt_grouped_sum}
\alias{tt_grouped_sum}
\title{grouped sum}
\usage{
tt_grouped_sum(data, ...)
}
\arguments{
\item{data}{data frame}

\item{...}{grouped columns}
}
\value{
data frame
}
\description{
grouped sum
}
