% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industry_grouped_sum.R
\name{tt_industry_grouped_sum}
\alias{tt_industry_grouped_sum}
\title{tt_industry_grouped_sum}
\usage{
tt_industry_grouped_sum(x, industry_type = "all_industry", sub = 11,
  verbose = TRUE)
}
\arguments{
\item{x}{data.frame}

\item{industry_type}{"all_industry", "industry21", "version1", "version2"}

\item{sub}{sub hscode}

\item{verbose}{TRUE or FALSE}
}
\value{
a list
}
\description{
tt_industry_grouped_sum
}
