% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mof.R
\name{tt_read_mof}
\alias{tt_read_mof}
\title{Read Data from MOF SOURCE}
\usage{
tt_read_mof(start_date, end_date, period = 0, direct = "export",
  money = "usd", columns = NULL)
}
\arguments{
\item{start_date}{Start date with "year-month" format. ex> "2019-01".}

\item{end_date}{End date with "year-month" format. ex> "2019-01".}

\item{period}{Integer}

\item{direct}{A string used to specify `export` and `import` values. The default value is `export`}

\item{money}{A string used to specify `usd` and `twd` value. The default value is `usd`}

\item{columns}{a character vector}
}
\value{
data.frame
}
\description{
Reading mof data with `start date` and `end date`. If you want to read data with many year you can specify the `period`.
}
